Moving files around and making sure all imports are as intended
