+++ src/App.tsx
@@ -1,8 +1,8 @@
 import React, { useState } from 'react';
 import './App.css';
 import { Button, Form } from 'react-bootstrap';
-import Navbar2 from './Components/Navbar'
-import { HomeScreen } from './HomeScreen/HomeScreen';
+import Navbar2 from './Components/Navbar/Navbar'
+import { HomeScreen } from './Components/HomeScreen/HomeScreen';
 
 //local storage and API Key: key should be entered in by the user and will be stored in local storage (NOT session storage)
 let keyData = "";
