+++ tsconfig.json
@@ -22,5 +22,5 @@
   },
   "include": [
     "src"
-, "src/HomeScreen/HomeScreen.tsx"  ]
+, "src/Components/HomeScreen/HomeScreen.tsx"  ]
 }
